REFACTOR:  rename `NumpyArray` to `Array`, fix tests
